deployment: merges test transfer script
